Add configurable sleeping delay between knocks
This commit changes the default behavior by introducing a sleeping delay
between connection attempts, if the number of ports are more than one.
The default value for this delay has been set to one (1.0) second.

The delay setting is configurable using a new command line parameter
(-w or --wait).

This closes #1 (Add flag for custom period to sleep between knocks).